requirements.txt
# Chrono AI Assignment Grader - Requirements
# Install with: pip install -r requirements.txt

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.0.0

# OpenAI API
openai==0.28.1

# Additional utilities
python-dotenv==1.0.0

# Optional: For PDF support
PyMuPDF==1.23.0
